// src/game/content/battle-sprite-atlases.generated.ts
export const BATTLE_SPRITE_ATLASES = [
  {
    "id": "map-combat",
    "textureKey": "battle-sprite-map-combat",
    "image": "/assets/original/battle-sprite-atlases/map-combat.png",
    "data": "/assets/original/battle-sprite-atlases/map-combat.json",
    "width": 1010,
    "height": 94,
    "decodedBytes": 379760,
    "frameCount": 46
  },
  {
    "id": "turn-transition",
    "textureKey": "battle-sprite-turn-transition",
    "image": "/assets/original/battle-sprite-atlases/turn-transition.png",
    "data": "/assets/original/battle-sprite-atlases/turn-transition.json",
    "width": 668,
    "height": 103,
    "decodedBytes": 275216,
    "frameCount": 9
  },
  {
    "id": "stage4-force-field-pulse",
    "textureKey": "battle-sprite-stage4-force-field-pulse",
    "image": "/assets/original/battle-sprite-atlases/stage4-force-field-pulse.png",
    "data": "/assets/original/battle-sprite-atlases/stage4-force-field-pulse.json",
    "width": 540,
    "height": 48,
    "decodedBytes": 103680,
    "frameCount": 13
  },
  {
    "id": "stage26-column-push",
    "textureKey": "battle-sprite-stage26-column-push",
    "image": "/assets/original/battle-sprite-atlases/stage26-column-push.png",
    "data": "/assets/original/battle-sprite-atlases/stage26-column-push.json",
    "width": 1010,
    "height": 94,
    "decodedBytes": 379760,
    "frameCount": 41
  }
] as const;

export const BATTLE_SPRITE_ATLAS_FRAMES = {
// ...
  "/assets/original/map-combat/hit/00.png": {
    "atlas": "map-combat",
    "frame": "map-combat/hit/00"
  },
// ...
  "/assets/original/stage26-column-push/phase2/10.png": {
    "atlas": "stage26-column-push",
    "frame": "stage26-column-push/phase2/10"
  }
} as const;